Restructuring vmec adjoint stuff.
