+++ src/simsopt/mhd/vmec.py
@@ -33,7 +33,6 @@ except ImportError as e:
 from .._core.optimizable import Optimizable
 from .._core.util import Struct, ObjectiveFailure
 from ..geo.surfacerzfourier import SurfaceRZFourier
-from ..geo.surfaceobjectives import parameter_derivatives
 
 if MPI is not None:
     from ..util.mpi import MpiPartition
@@ -595,353 +594,3 @@ class Vmec(Optimizable):
 
         well = (dVds_s0 - dVds_s1) / dVds_s0
         return well
-
-    def iota_weighted(self, weight_function):
-        """
-        Computed a weighted average of the rotational transform defined by
-        the prescribed weight_function, f.
-
-        \int ds \, iota * weight_function(s) / \int ds \, weight_function(s)
-
-        Args:
-            weight_function : function handle which takes a single argument, s,
-                the normalized toroidal flux
-        """
-        self.run()
-
-        return np.sum(weight_function(self.s_half_grid) * self.wout.iotas[1:]) \
-            / np.sum(weight_function(self.s_half_grid))
-
-    def iota_target_metric(self, iota_target):
-        """
-        Computes a metric quantifying the deviation of iota from a prescribed
-        target value.
-
-        0.5 * \int ds \, (iota - iota_target)**2
-
-        Args:
-            iota_target : function handle which takes a single argument, s,
-                the normalized toroidal flux, and returns the target rotational
-                transform.
-        """
-        self.run()
-        return 0.5 * np.sum((self.wout.iotas[1::]
-                             - iota_target(self.s_half_grid))**2) * self.ds
-
-    def d_iota_target_metric(self, iota_target, delta=1.):
-        """
-        Computes derivatives of iota_target_metric wrt surface parameters using
-        an adjoint method. The parameter delta sets the amplitude of the toroidal
-        current perturbation required for the adjoint solve.
-        """
-        if self.indata.ncurr != 1:
-            raise RuntimeError('''d_iota_target_metric cannot be computed without
-                running vmec with ncurr = 1''')
-
-        shape_gradient = self.iota_target_metric_shape_gradient(iota_target, delta)
-
-        return parameter_derivatives(self.boundary, shape_gradient)
-
-    def iota_target_metric_shape_gradient(self, iota_target, delta):
-        """
-        Computes shape gradient of iota_target_metric, defined as S where,
-
-        \delta f(\delta x) = \int d^2 x \, S \delta x \cdot n,
-
-        is the perturbation to the objective function corresponding to the
-        perturbation of the surface, \delta x.
-        """
-        self.run()
-
-        Bx0, By0, Bz0 = self.B_cartesian()
-
-        mu0 = 4*np.pi*1e-7
-        It_half = self.wout.signgs * 2*np.pi * self.wout.bsubumnc[0, 1::] / mu0
-        ac_aux_f_prev = np.copy(self.indata.ac_aux_f)
-        ac_aux_s_prev = np.copy(self.indata.ac_aux_s)
-        pcurr_type_prev = np.copy(self.indata.pcurr_type)
-        curtor_prev = np.copy(self.indata.curtor)
-
-        perturbation = (self.wout.iotas[1::]-iota_target(self.s_half_grid)) \
-            / (self.wout.phi[-1]*self.wout.signgs/(2*np.pi))
-
-        # Perturbed toroidal current profile
-        It_new = It_half + delta*perturbation
-        curtor = 1.5*It_new[-1] - 0.5*It_new[-2]
-        self.indata.ac_aux_f = -1.*np.ones_like(self.indata.ac_aux_f)
-        self.indata.ac_aux_s = -1.*np.ones_like(self.indata.ac_aux_s)
-        self.indata.ac_aux_f[0:self.wout.ns-1] = It_new
-        self.indata.ac_aux_s[0:self.wout.ns-1] = self.s_half_grid
-        self.indata.curtor = curtor
-        self.indata.pcurr_type = b'line_segment_I'
-        self.need_to_run_code = True
-
-        self.run()
-
-        It_half = self.wout.signgs * 2*np.pi * self.wout.bsubumnc[0, 1::] / mu0
-
-        Bx, By, Bz = self.B_cartesian()
-
-        # Reset input values
-        self.indata.ac_aux_f = ac_aux_f_prev
-        self.indata.ac_aux_s = ac_aux_s_prev
-        self.indata.pcurr_type = pcurr_type_prev
-        self.indata.curtor = curtor_prev
-        self.need_to_run_code = True
-
-        deltaB_dot_B = ((Bx-Bx0)*Bx0 + (By-By0)*By0 + (Bz-Bz0)*Bz0)/delta
-
-        return deltaB_dot_B/(2*np.pi*mu0)
-
-    def iota_weighted_shape_gradient(self, weight_function, delta):
-        """
-        Computes shape gradient of iota_weighted, defined as S where,
-
-        \delta f(\delta x) = \int d^2 x \, S \delta x \cdot n,
-
-        is the perturbation to the objective function corresponding to the
-        perturbation of the surface, \delta x.
-        """
-        self.run()
-
-        Bx0, By0, Bz0 = self.B_cartesian()
-
-        mu0 = 4*np.pi*1e-7
-        It_half = self.wout.signgs * 2*np.pi * self.wout.bsubumnc[0, 1::] / mu0
-        ac_aux_f_prev = np.copy(self.indata.ac_aux_f)
-        ac_aux_s_prev = np.copy(self.indata.ac_aux_s)
-        pcurr_type_prev = np.copy(self.indata.pcurr_type)
-        curtor_prev = np.copy(self.indata.curtor)
-
-        perturbation = (self.wout.iotas[1::]-iota_target(self.s_half_grid)) \
-            / (self.wout.phi[-1]*self.wout.signgs/(2*np.pi))
-
-        # Perturbed toroidal current profile
-        It_new = It_half + delta*perturbation
-        curtor = 1.5*It_new[-1] - 0.5*It_new[-2]
-        self.indata.ac_aux_f = -1.*np.ones_like(self.indata.ac_aux_f)
-        self.indata.ac_aux_s = -1.*np.ones_like(self.indata.ac_aux_s)
-        self.indata.ac_aux_f[0:self.wout.ns-1] = It_new
-        self.indata.ac_aux_s[0:self.wout.ns-1] = self.s_half_grid
-        self.indata.curtor = curtor
-        self.indata.pcurr_type = b'line_segment_I'
-        self.need_to_run_code = True
-
-        self.run()
-
-        It_half = self.wout.signgs * 2*np.pi * self.wout.bsubumnc[0, 1::] / mu0
-
-        Bx, By, Bz = self.B_cartesian()
-
-        # Reset input values
-        self.indata.ac_aux_f = ac_aux_f_prev
-        self.indata.ac_aux_s = ac_aux_s_prev
-        self.indata.pcurr_type = pcurr_type_prev
-        self.indata.curtor = curtor_prev
-        self.need_to_run_code = True
-
-        deltaB_dot_B = ((Bx-Bx0)*Bx0 + (By-By0)*By0 + (Bz-Bz0)*Bz0)/delta
-
-        return deltaB_dot_B/(2*np.pi*mu0)
-
-    def d_well_weighted(self, weight_function1, weight_function2, delta=1.):
-        """
-        Computes derivatives of well_weighted wrt surface parameters using
-        an adjoint method. The parameter delta sets the amplitude of
-        the toroidal current perturbation required for the adjoint
-        solve.
-        """
-
-        shape_gradient = self.well_weighted_shape_gradient(weight_function1, weight_function2, delta)
-
-        return parameter_derivatives(self.boundary, shape_gradient)
-
-    def well_weighted_shape_gradient(self, weight_function1, weight_function2, delta=1.):
-        """
-        Computes shape gradient of well_weighted_metric, defined as S where,
-
-        \delta f(\delta x) = \int d^2 x \, S \delta x \cdot n,
-
-        is the perturbation to the objective function corresponding to the
-        perturbation of the surface, \delta x.
-        """
-        self.run()
-
-        Bx0, By0, Bz0 = self.B_cartesian()
-
-        mu0 = 4*np.pi*1e-7
-        am_aux_f_prev = np.copy(self.indata.am_aux_f)
-        am_aux_s_prev = np.copy(self.indata.am_aux_s)
-        pmass_type_prev = np.copy(self.indata.pmass_type)
-
-        pres = self.wout.pres[1::]
-        weight1 = weight_function1(self.s_half_grid) - weight_function2(self.s_half_grid)
-        weight2 = weight_function1(self.s_half_grid) + weight_function2(self.s_half_grid)
-        numerator = np.sum(weight1 * self.wout.vp[1::])
-        denominator = np.sum(weight2 * self.wout.vp[1::])
-        fW = numerator/denominator
-        perturbation = (weight1 - fW * weight2) / (denominator * self.ds * 4 * np.pi * np.pi)
-
-        # Perturbed pressure profile
-        pres_new = pres + delta*perturbation
-
-        self.indata.am_aux_f = -1.*np.ones_like(self.indata.am_aux_f)
-        self.indata.am_aux_s = -1.*np.ones_like(self.indata.am_aux_s)
-        self.indata.am_aux_f[0:self.wout.ns-1] = pres_new
-        self.indata.am_aux_s[0:self.wout.ns-1] = self.s_half_grid
-        self.indata.pmass_type = b'cubic_spline'
-        self.need_to_run_code = True
-
-        self.run()
-
-        Bx, By, Bz = self.B_cartesian()
-
-        # Reset input values
-        self.indata.am_aux_f = am_aux_f_prev
-        self.indata.am_aux_s = am_aux_s_prev
-        self.indata.pmass_type = pmass_type_prev
-        self.need_to_run_code = True
-
-        deltaB_dot_B = ((Bx-Bx0)*Bx0 + (By-By0)*By0 + (Bz-Bz0)*Bz0)/delta
-
-        return deltaB_dot_B/(mu0) + perturbation[-1]
-
-    def B_cartesian(self):
-        """
-        Computes cartesian vector components of magnetic field on boundary
-        on a grid in the vmec toroidal and poloidal angles. This is
-        required to compute adjoint-based shape gradients.
-        """
-        dgamma1 = self.boundary.gammadash1()
-        dgamma2 = self.boundary.gammadash2()
-
-        theta1D = self.boundary.quadpoints_theta * 2 * np.pi
-        phi1D = self.boundary.quadpoints_phi * 2 * np.pi
-        nphi = len(phi1D)
-        ntheta = len(theta1D)
-        theta, phi = np.meshgrid(theta1D, phi1D)
-
-        self.run()
-        bsupumnc = 1.5 * self.wout.bsupumnc[:, -1] - 0.5 * self.wout.bsupumnc[:, -2]
-        bsupvmnc = 1.5 * self.wout.bsupvmnc[:, -1] - 0.5 * self.wout.bsupvmnc[:, -2]
-        angle = self.wout.xm_nyq[:, None, None] * theta[None, :, :] \
-            - self.wout.xn_nyq[:, None, None] * phi[None, :, :]
-        Bsupu = np.sum(bsupumnc[:, None, None] * np.cos(angle), axis=0)
-        Bsupv = np.sum(bsupvmnc[:, None, None] * np.cos(angle), axis=0)
-
-        Bx = (Bsupv * dgamma1[:, :, 0] + Bsupu * dgamma2[:, :, 0])/(2*np.pi)
-        By = (Bsupv * dgamma1[:, :, 1] + Bsupu * dgamma2[:, :, 1])/(2*np.pi)
-        Bz = (Bsupv * dgamma1[:, :, 2] + Bsupu * dgamma2[:, :, 2])/(2*np.pi)
-
-        return Bx, By, Bz
-
-    def arclength_poloidal_angle(self):
-        """
-        Computes poloidal angle based on arclenth along magnetic surface at
-        constant phi. This is required for evaluating the adjoint shape gradient
-        for free-boundary calculations.
-        """
-        gamma = self.boundary.gamma()
-        X = gamma[:, :, 0]
-        Y = gamma[:, :, 1]
-        Z = gamma[:, :, 2]
-        R = np.sqrt(X**2 + Y**2)
-
-        theta_arclength = np.zeros_like(gamma[:, :, 0])
-        nphi = len(theta_arclength[:, 0])
-        ntheta = len(theta_arclength[0, :])
-        for iphi in range(nphi):
-            for itheta in range(1, ntheta):
-                dr = np.sqrt((R[iphi, itheta] - R[iphi, itheta-1])**2
-                             + (Z[iphi, itheta] - Z[iphi, itheta-1])**2)
-                theta_arclength[iphi, itheta] = \
-                    theta_arclength[iphi, itheta-1] + dr
-        return theta_arclength
-
-    def interpolate_on_arclength_grid(self, function, theta_evaluate):
-        """
-        Interpolate function onto the theta_evaluate grid in the arclength
-        poloidal angle.
-        """
-        from scipy import interpolate
-
-        theta_arclength = self.arclength_poloidal_angle()
-        function_interpolated = np.zeros_like(function)
-        nphi = len(theta_arclength[:, 0])
-        for iphi in range(nphi):
-            f = interpolate.InterpolatedUnivariateSpline(
-                theta_arclength[iphi, :], function[iphi, :])
-            function_interpolated[iphi, :] = f(theta_evaluate[iphi, :])
-
-        return function_interpolated
-
-    def well_weighted(self, weight_function1, weight_function2):
-        """
-        Computed a weighted average of the vacuum magnetic well defined by
-        the prescribed weight_functions:
-
-        f = \int ds \, V'(s) * (weight_function1(s) - weight_function2(s))
-            / \int ds \, V'(s) (weight_function1(s) + weight_function2(s))
-
-        For example, weight_function1 could be peaked on the edge while
-        weight_function2 could be peaked on the axis such that f < 0 corresonds
-        to V''(s) < 0.
-
-        Args:
-            weight_function1 : function handle which takes a single argument, s,
-                the normalized toroidal flux
-            weight_function2 : function handle which takes a single argument, s,
-                the normalized toroidal flux
-        """
-        self.run()
-        return np.sum((weight_function1(self.s_half_grid)-weight_function2(self.s_half_grid)) * self.wout.vp[1:]) \
-            / np.sum((weight_function1(self.s_half_grid)+weight_function2(self.s_half_grid)) * self.wout.vp[1:])
-
-
-class IotaTargetMetric(Optimizable):
-    """
-    IotaTargetMetric is a class that computes iota_target_metric from a
-    vmec equilibrium. Its derivatives can also be computed with an adjoint
-    method.
-    """
-
-    def __init__(self, vmec, iota_target, adjoint_epsilon=1.e-1):
-        self.vmec = vmec
-        self.boundary = vmec.boundary
-        self.iota_target = iota_target
-        self.adjoint_epsilon = adjoint_epsilon
-        self.depends_on = ["boundary"]
-
-    def J(self):
-        self.vmec.need_to_run_code = True
-        return self.vmec.iota_target_metric(self.iota_target)
-
-    def dJ(self):
-        self.vmec.need_to_run_code = True
-        grad = self.vmec.d_iota_target_metric(self.iota_target, self.adjoint_epsilon)
-        return grad
-
-
-class WellWeighted(Optimizable):
-    """
-    WellWeighted is a class that computes well_weighted from a vmec
-    equilibrium. Its derivatives can also be computed with an adjoint
-    method.
-    """
-
-    def __init__(self, vmec, weight_function1, weight_function2, adjoint_epsilon=1.e-1):
-        self.vmec = vmec
-        self.boundary = vmec.boundary
-        self.weight_function1 = weight_function1
-        self.weight_function2 = weight_function2
-        self.adjoint_epsilon = adjoint_epsilon
-        self.depends_on = ["boundary"]
-
-    def J(self):
-        self.vmec.need_to_run_code = True
-        return self.vmec.well_weighted(self.weight_function1, self.weight_function2)
-
-    def dJ(self):
-        self.vmec.need_to_run_code = True
-        grad = self.vmec.d_well_weighted(self.weight_function1, self.weight_function2, self.adjoint_epsilon)
-        return grad
