+++ tests/geo/test_surface.py
@@ -1,6 +1,7 @@
 import unittest
 from pathlib import Path
 import numpy as np
+import os
 
 from simsopt._core.dofs import Dofs
 from simsopt._core.optimizable import make_optimizable
@@ -105,7 +106,7 @@ class SurfaceXYZFourierTests(unittest.TestCase):
         Away from the quadrature points, the conversion is not lossless and this test verifies that the
         error is small.
 
-        Additionally, this test checks that the cross sectional angle is correct.        
+        Additionally, this test checks that the cross sectional angle is correct.
         """
         s = get_exact_surface()
         sRZ = s.to_RZFourier()
@@ -138,7 +139,7 @@ class SurfaceXYZFourierTests(unittest.TestCase):
     def test_cross_section_torus(self):
         """
         Test that the cross sectional area at a certain number of cross sections of a torus
-        is what it should be.  The cross sectional angles are chosen to test any degenerate 
+        is what it should be.  The cross sectional angles are chosen to test any degenerate
         cases of the bisection algorithm.
 
         Additionally, this test checks that the cross sectional angle is correct.
@@ -247,7 +248,7 @@ class SurfaceXYZFourierTests(unittest.TestCase):
 
     def test_aspect_ratio_compare_with_cross_sectional_computation(self):
         """
-        This test validates the VMEC aspect ratio computation in the Surface class by 
+        This test validates the VMEC aspect ratio computation in the Surface class by
         comparing with an approximation based on cross section computations.
         """
         s = get_exact_surface()
@@ -263,7 +264,7 @@ class SurfaceXYZFourierTests(unittest.TestCase):
             Z = cs[:, 2]
             Rp = fftpack.diff(R, period=1.)
             Zp = fftpack.diff(Z, period=1.)
-            ar = np.mean(Z*Rp) 
+            ar = np.mean(Z*Rp)
             cs_area[idx] = ar
 
         mean_cross_sectional_area = np.mean(cs_area)
@@ -356,8 +357,8 @@ class SurfaceRZFourierTests(unittest.TestCase):
 
         # Now try a nonaxisymmetric shape:
         s = SurfaceRZFourier(mpol=3, ntor=1)
-        s.rc[:, :] = [[100, 2, 3], [4, 5, 6], [7, 8, 9], [10, 11, 12]] 
-        s.zs[:, :] = [[101, 102, 13], [14, 15, 16], [17, 18, 19], [20, 21, 22]] 
+        s.rc[:, :] = [[100, 2, 3], [4, 5, 6], [7, 8, 9], [10, 11, 12]]
+        s.zs[:, :] = [[101, 102, 13], [14, 15, 16], [17, 18, 19], [20, 21, 22]]
         dofs = s.get_dofs()
         self.assertEqual(dofs.shape, (21,))
         for j in range(21):
@@ -556,6 +557,59 @@ class SurfaceGarabedianTests(unittest.TestCase):
                     sf2 = sg.to_RZFourier()
                     np.testing.assert_allclose(sf1.rc, sf2.rc)
                     np.testing.assert_allclose(sf1.zs, sf2.zs)
+
+
+class ArclengthTests(unittest.TestCase):
+    def test_arclength_poloidal_angle(self):
+        """
+        Compute arclength poloidal angle from circular cross-section tokamak.
+        Check that this matches parameterization angle.
+        """
+        s = get_surface('SurfaceRZFourier', True, mpol=1, ntor=0,
+                        ntheta=200, nphi=5, full=True)
+        s.rc[0, 0] = 5.
+        s.rc[1, 0] = 1.
+        s.zs[1, 0] = 1.
+
+        theta1D = s.quadpoints_theta * 2 * np.pi
+
+        arclength = s.arclength_poloidal_angle()
+        nphi = len(arclength[:, 0])
+        for iphi in range(nphi):
+            self.assertTrue(np.max(np.abs(arclength[iphi, :]-theta1D)) < 1e-3)
+
+    def test_interpolate_on_arclength_grid(self):
+        """
+        Check that line integral of (1 + cos(theta - phi)) at constant phi is
+        unchanged when evaluated on parameterization or arclength poloidal angle
+        grid.
+        """
+        ntheta = 500
+        nphi = 10
+        s = get_surface('SurfaceRZFourier', True, mpol=5, ntor=5,
+                        ntheta=ntheta, nphi=nphi, full=True)
+        s.rc[0, 0] = 5.
+        s.rc[1, 0] = -1.5
+        s.zs[1, 0] = -1.5
+        s.rc[1, 1] = -0.5
+        s.zs[1, 1] = 0.5
+        s.rc[0, 1] = -0.5
+        s.zs[0, 1] = 0.5
+
+        dgamma2 = s.gammadash2()
+        theta1D = s.quadpoints_theta
+        phi1D = s.quadpoints_phi
+        theta, phi = np.meshgrid(theta1D, phi1D)
+        integrand = 1 + np.cos(theta - phi)
+
+        norm_drdtheta = np.linalg.norm(dgamma2, axis=2)
+        length = np.sum(norm_drdtheta, axis=1) / ntheta
+        theta_interp = theta * length[:, None]
+        integrand_arclength = s.interpolate_on_arclength_grid(integrand, theta_interp)
+        for iphi in range(nphi):
+            integral_1 = np.sum(integrand[iphi, :] * norm_drdtheta[iphi, :]) / np.sum(norm_drdtheta[iphi, :])
+            integral_2 = np.sum(integrand_arclength[iphi, :]) / np.sum(np.ones_like(norm_drdtheta[iphi, :]))
+            self.assertAlmostEqual(integral_1, integral_2, places=3)
 
 
 if __name__ == "__main__":
