+++ tests/mhd/test_vmec.py
@@ -2,7 +2,6 @@ import unittest
 import logging
 import os
 import numpy as np
-from simsopt.mhd.vmec import IotaTargetMetric, WellWeighted
 
 try:
     from mpi4py import MPI
@@ -193,327 +192,6 @@ class VmecTests(unittest.TestCase):
 
         self.assertAlmostEqual(mean_iota, mean_iota_alt, places=3)
         self.assertAlmostEqual(mean_shear, mean_shear_alt, places=3)
-
-    def test_iota_target_metric(self):
-        """
-        Check that iota_target_metric computed on full instead of half grid
-        yields approximately same result.
-        """
-        filename = os.path.join(TEST_DIR, 'input.rotating_ellipse')
-        vmec = Vmec(filename, ntheta=50, nphi=50)
-
-        target_function = lambda s: np.cos(s)
-        metric1 = vmec.iota_target_metric(target_function)
-        stencil = np.ones_like(vmec.s_full_grid)
-        stencil[0] = 0.5
-        stencil[-1] = 0.5
-        metric2 = 0.5 * np.sum((vmec.wout.iotaf - target_function(vmec.s_full_grid))**2 * stencil) * vmec.ds
-
-        self.assertAlmostEqual(metric1, metric2, places=3)
-
-    def test_iota_weighted(self):
-        """
-        Check that objective value with peaked gaussian weight function yields
-        approximately the value of iota at center of weight function.
-        """
-        filename = os.path.join(TEST_DIR, 'input.rotating_ellipse')
-        vmec = Vmec(filename, ntheta=50, nphi=50)
-
-        vmec.run()
-        iota_center = vmec.wout.iotas[50]
-        s_center = vmec.s_half_grid[50]
-        target_function = lambda s: np.exp(-(s-s_center)**2/0.01**2)
-        self.assertAlmostEqual(vmec.iota_weighted(target_function), iota_center, places=2)
-
-    def test_well_weighted(self):
-        """
-        Check that objective value with peaked gaussian weight function at axis
-        and edge matches well metric computed with V'(0) and V'(1).
-        """
-        filename = os.path.join(TEST_DIR, 'input.rotating_ellipse')
-        vmec = Vmec(filename, ntheta=50, nphi=50)
-
-        vmec.run()
-        vp_l = 1.5*vmec.wout.vp[1]-0.5*vmec.wout.vp[2]
-        vp_r = 1.5*vmec.wout.vp[-1]-0.5*vmec.wout.vp[-2]
-        well1 = (vp_l-vp_r)/(vp_l+vp_r)
-        weight1 = lambda s: np.exp(-s**2/0.01**2)
-        weight2 = lambda s: np.exp(-(1-s)**2/0.01**2)
-        well2 = vmec.well_weighted(weight1, weight2)
-        self.assertAlmostEqual(well1, well2, places=2)
-
-    def test_B_cartesian(self):
-        """
-        Check that B^2 matches bmnc from wout file.
-        """
-        filename = os.path.join(TEST_DIR, 'input.rotating_ellipse')
-        vmec = Vmec(filename, ntheta=50, nphi=50)
-
-        Bx, By, Bz = vmec.B_cartesian()
-        B2 = Bx*Bx + By*By + Bz*Bz
-
-        theta1D = vmec.boundary.quadpoints_theta * 2 * np.pi
-        phi1D = vmec.boundary.quadpoints_phi * 2 * np.pi
-        nphi = len(phi1D)
-        ntheta = len(theta1D)
-        theta, phi = np.meshgrid(theta1D, phi1D)
-
-        bmnc = 1.5 * vmec.wout.bmnc[:, -1] - 0.5 * vmec.wout.bmnc[:, -2]
-        xm = vmec.wout.xm_nyq
-        xn = vmec.wout.xn_nyq
-        angle = vmec.wout.xm_nyq[:, None, None] * theta[None, :, :] \
-            - vmec.wout.xn_nyq[:, None, None] * phi[None, :, :]
-        B = np.sum(bmnc[:, None, None] * np.cos(angle), axis=0)
-        self.assertTrue(np.max(np.abs(B**2-B2)) < 1.e-4)
-
-    def test_arclength_poloidal_angle(self):
-        """
-        Compute arclength poloidal angle from circular cross-section tokamak.
-        Check that this matches parameterization angle.
-        """
-        filename = os.path.join(TEST_DIR, 'input.tok')
-        vmec = Vmec(filename, ntheta=500, nphi=5)
-        theta1D = vmec.boundary.quadpoints_theta * 2 * np.pi
-
-        arclength = vmec.arclength_poloidal_angle()
-        nphi = len(arclength[:, 0])
-        for iphi in range(nphi):
-            self.assertTrue(np.max(np.abs(arclength[iphi, :]-theta1D)) < 1e-3)
-
-    def test_interpolate_on_arclength_grid(self):
-        """
-        Check that line integral of B^2 at constant phi is unchanged when
-        evaluated on parameterization or arclength poloidal angle grid.
-        """
-        filename = os.path.join(TEST_DIR, 'input.rotating_ellipse')
-        ntheta = 500
-        nphi = 10
-        vmec = Vmec(filename, ntheta=ntheta, nphi=nphi)
-
-        Bx, By, Bz = vmec.B_cartesian()
-        B2 = Bx*Bx + By*By + Bz*Bz
-        dgamma2 = vmec.boundary.gammadash2()
-
-        theta1D = vmec.boundary.quadpoints_theta
-        phi1D = vmec.boundary.quadpoints_phi
-        theta, phi = np.meshgrid(theta1D, phi1D)
-
-        norm_drdtheta = np.linalg.norm(dgamma2, axis=2)
-        length = np.sum(norm_drdtheta, axis=1) / ntheta
-        theta_interp = theta * length[:, None]
-        B2_arclength = vmec.interpolate_on_arclength_grid(B2, theta_interp)
-        for iphi in range(nphi):
-            integral_1 = np.sum(B2[iphi, :] * norm_drdtheta[iphi, :]) / np.sum(norm_drdtheta[iphi, :])
-            integral_2 = np.sum(B2_arclength[iphi, :]) / np.sum(np.ones_like(norm_drdtheta[iphi, :]))
-            self.assertAlmostEqual(integral_1, integral_2, places=3)
-
-    def test_d_iota_target_metric(self):
-        """
-        Assert that RuntimeError is raised if ncurr = 0.
-        Compare d_iota_target_metric with finite differences for a surface
-        perturbation in a random direction.
-        """
-        filename = os.path.join(TEST_DIR, 'input.rotating_ellipse_refined')
-        vmec = Vmec(filename, ntheta=100, nphi=100)
-
-        target_function = lambda s: 0.68
-        epsilon = 1.e-4  # FD step size
-        adjoint_epsilon = 1.e-1  # perturbation amplitude for adjoint solve
-
-        # Check that assert raised if ncurr = 0
-        vmec.indata.ncurr = 0
-        with self.assertRaises(RuntimeError):
-            vmec.d_iota_target_metric(target_function)
-        vmec.indata.ncurr = 1
-
-        # Compute random direction for surface perturbation
-        surf = vmec.boundary
-        dofs = np.copy(vmec.boundary.get_dofs())
-        vec = np.random.standard_normal(dofs.shape)
-        unitvec = vec / np.sqrt(np.vdot(vec, vec))
-
-        def iota_fun(epsilon):
-            vmec.boundary.set_dofs(dofs + epsilon*unitvec)
-            vmec.need_to_run_code = True
-            return vmec.iota_target_metric(target_function)
-
-        d_iota_fd = (iota_fun(epsilon)-iota_fun(-epsilon))/(2*epsilon)
-
-        vmec.boundary.set_dofs(dofs)
-        vmec.need_to_run_code = True
-        d_iota_adjoint = np.dot(vmec.d_iota_target_metric(target_function, adjoint_epsilon), unitvec)
-
-        relative_error = np.abs(d_iota_fd-d_iota_adjoint)/np.abs(d_iota_fd)
-        logger.info(f"adjoint jac: {d_iota_adjoint},   fd jac: {d_iota_fd}")
-        logger.info(f"relative error: {relative_error}")
-        self.assertTrue(relative_error < 5.e-2)
-
-    def test_d_well_weighted(self):
-        """
-        Compare d_well_weighted_metric with finite differences for a surface
-        perturbation in a random direction.
-        """
-        filename = os.path.join(TEST_DIR, 'input.rotating_ellipse_refined')
-        vmec = Vmec(filename, ntheta=100, nphi=100)
-
-        epsilon = 1.e-2  # FD step size
-        adjoint_epsilon = 1.e-1  # perturbation amplitude for adjoint solve
-        weight1 = lambda s: np.exp(-s**2/0.2**2)
-        weight2 = lambda s: np.exp(-(1-s)**2/0.2**2)
-
-        # Compute random direction for surface perturbation
-        surf = vmec.boundary
-        dofs = np.copy(vmec.boundary.get_dofs())
-        vec = np.random.standard_normal(dofs.shape)
-        unitvec = vec / np.sqrt(np.vdot(vec, vec))
-
-        def well_fun(epsilon):
-            vmec.boundary.set_dofs(dofs + epsilon*unitvec)
-            vmec.need_to_run_code = True
-            return vmec.well_weighted(weight1, weight2)
-
-        d_well_fd = (well_fun(epsilon)-well_fun(-epsilon))/(2*epsilon)
-
-        vmec.boundary.set_dofs(dofs)
-        vmec.need_to_run_code = True
-        d_well_adjoint = np.dot(vmec.d_well_weighted(weight1, weight2, adjoint_epsilon), unitvec)
-
-        relative_error = np.abs(d_well_fd-d_well_adjoint)/np.abs(d_well_fd)
-        logger.info(f"adjoint jac: {d_well_adjoint},   fd jac: {d_well_fd}")
-        logger.info(f"relative error: {relative_error}")
-        self.assertTrue(relative_error < 5.e-2)
-
-    def test_WellWeighted(self):
-        """
-        Compare dJ() with finite differences for a surface perturbation in a
-        random direction.
-        """
-        filename = os.path.join(TEST_DIR, 'input.rotating_ellipse')
-        vmec = Vmec(filename, ntheta=100, nphi=100)
-
-        epsilon = 1.e-2  # FD step size
-        adjoint_epsilon = 1.e-1  # perturbation amplitude for adjoint solve
-
-        weight1 = lambda s: np.exp(-s**2/0.2**2)
-        weight2 = lambda s: np.exp(-(1-s)**2/0.2**2)
-
-        obj = WellWeighted(vmec, weight1, weight2, adjoint_epsilon)
-
-        # Compute random direction for surface perturbation
-        dofs = np.copy(vmec.boundary.get_dofs())
-        vec = np.random.standard_normal(dofs.shape)
-        unitvec = vec / np.sqrt(np.vdot(vec, vec))
-
-        def well_fun(epsilon):
-            vmec.boundary.set_dofs(dofs + epsilon*unitvec)
-            return obj.J()
-
-        d_well_fd = (well_fun(epsilon)-well_fun(-epsilon))/(2*epsilon)
-
-        vmec.boundary.set_dofs(dofs)
-        vmec.need_to_run_code = True
-        d_well_adjoint = np.dot(obj.dJ(), unitvec)
-
-        relative_error = np.abs(d_well_fd-d_well_adjoint)/np.abs(d_well_fd)
-        logger.info(f"adjoint jac: {d_well_adjoint},   fd jac: {d_well_fd}")
-        logger.info(f"relative error: {relative_error}")
-        self.assertTrue(relative_error < 5.e-2)
-
-    def test_WellWeighted_LeastSquaresProblem(self):
-        """
-        Compare jacobian for least-squares problem with supplied gradient wrt
-        one surface coefficient matches finite-differences.
-        """
-        filename = os.path.join(TEST_DIR, 'input.rotating_ellipse')
-        vmec = Vmec(filename, ntheta=100, nphi=100)
-
-        epsilon = 1.e-2  # FD step size
-        adjoint_epsilon = 1.e0  # perturbation amplitude for adjoint solve
-
-        weight1 = lambda s: np.exp(-s**2/0.4**2)
-        weight2 = lambda s: np.exp(-(1-s)**2/0.4**2)
-
-        # Compute random direction for surface perturbation
-        surf = vmec.boundary
-        surf.all_fixed()
-        surf.set_fixed("rc(0,0)", False)  # Major radius
-
-        obj = WellWeighted(vmec, weight1, weight2, adjoint_epsilon)
-
-        prob = LeastSquaresProblem([(obj, 0, 1)])
-        prob.dofs.abs_step = epsilon
-
-        jac = prob.dofs.jac()
-        fd_jac = prob.dofs.fd_jac()
-
-        relative_error = np.abs(fd_jac-jac)/np.abs(fd_jac)
-        logger.info(f"adjoint jac: {jac},   fd jac: {fd_jac}")
-        logger.info(f"relative error: {relative_error}")
-        self.assertTrue(relative_error < 1.5e-2)
-
-    def test_IotaTargetMetric(self):
-        """
-        Compare dJ() with finite differences for a surface perturbation in a
-        random direction.
-        """
-        filename = os.path.join(TEST_DIR, 'input.rotating_ellipse')
-        vmec = Vmec(filename, ntheta=100, nphi=100)
-
-        target_function = lambda s: 0.68
-        epsilon = 1.e-4  # FD step size
-        adjoint_epsilon = 1.e-1  # perturbation amplitude for adjoint solve
-
-        obj = IotaTargetMetric(vmec, target_function, adjoint_epsilon)
-
-        # Compute random direction for surface perturbation
-        dofs = np.copy(vmec.boundary.get_dofs())
-        vec = np.random.standard_normal(dofs.shape)
-        unitvec = vec / np.sqrt(np.vdot(vec, vec))
-
-        def iota_fun(epsilon):
-            vmec.boundary.set_dofs(dofs + epsilon*unitvec)
-            return obj.J()
-
-        d_iota_fd = (iota_fun(epsilon)-iota_fun(-epsilon))/(2*epsilon)
-
-        vmec.boundary.set_dofs(dofs)
-        vmec.need_to_run_code = True
-        d_iota_adjoint = np.dot(obj.dJ(), unitvec)
-
-        relative_error = np.abs(d_iota_fd-d_iota_adjoint)/np.abs(d_iota_fd)
-        logger.info(f"adjoint jac: {d_iota_adjoint},   fd jac: {d_iota_fd}")
-        logger.info(f"relative error: {relative_error}")
-        self.assertTrue(relative_error < 5.e-2)
-
-    def test_IotaTargetMetric_LeastSquaresProblem(self):
-        """
-        Compare jacobian for least-squares problem with supplied gradient wrt
-        one surface coefficient matches finite-differences.
-        """
-        filename = os.path.join(TEST_DIR, 'input.rotating_ellipse')
-        vmec = Vmec(filename, ntheta=100, nphi=100)
-
-        target_function = lambda s: 0.68
-        epsilon = 1.e-3  # FD step size
-        adjoint_epsilon = 1.e0  # perturbation amplitude for adjoint solve
-
-        # Compute random direction for surface perturbation
-        surf = vmec.boundary
-        surf.all_fixed()
-        surf.set_fixed("rc(0,0)", False)  # Major radius
-
-        obj = IotaTargetMetric(vmec, target_function, adjoint_epsilon)
-
-        prob = LeastSquaresProblem([(obj, 0, 1)])
-
-        prob.dofs.abs_step = epsilon
-        jac = prob.dofs.jac()
-        fd_jac = prob.dofs.fd_jac()
-
-        relative_error = np.abs(fd_jac-jac)/np.abs(fd_jac)
-        logger.info(f"adjoint jac: {jac},   fd jac: {fd_jac}")
-        logger.info(f"relative error: {relative_error}")
-        self.assertTrue(relative_error < 1.e-2)
 
     #def test_stellopt_scenarios_1DOF_circularCrossSection_varyR0_targetVolume(self):
         """
